104/108 AG, one more

+++ src/StaticRouter.h
@@ -39,7 +39,7 @@ private:
         uint8_t icmp_type;
         uint8_t icmp_code;
         uint16_t icmp_sum;
-        uint16_t unused;
+        uint32_t unused;
         uint8_t data[ICMP_DATA_SIZE];
 
     } __attribute__ ((packed)) ;
